fix(mediabuilder): keep static fields from db
for the UUID and "first seen date" on Media, we'll want to save those from the database instead of making new ones each time.

having them as fields on the `MediaBuilder` was a bad call, so now they're gone, to be generated if they don't exist already in the database.

this doesn't yet account for media files that were moved, but checking their hash should help with that later on!